PMA-127 - TowerPolicy

+++ app/Http/Controllers/TowerController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\Tower\UpdateRequest;
 use App\Http\Resources\TowerResource;
 use App\Http\Resources\TowerResourceCollection;
 use App\Repositories\Contracts\TowerRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class TowerController extends Controller
 {
@@ -32,9 +33,12 @@ class TowerController extends Controller
      *
      * @param IndexRequest $request
      * @return TowerResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [Tower::class, $request->get('propertyId')]);
+
         $towers = $this->towerRepository->findBy($request->all());
 
         return new TowerResourceCollection($towers);
@@ -45,9 +49,12 @@ class TowerController extends Controller
      *
      * @param  StoreRequest  $request
      * @return TowerResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [Tower::class, $request->get('propertyId')]);
+
         $tower = $this->towerRepository->save($request->all());
 
         return new TowerResource($tower);
@@ -58,9 +65,12 @@ class TowerController extends Controller
      *
      * @param  Tower  $tower
      * @return TowerResource
+     * @throws AuthorizationException
      */
     public function show(Tower $tower)
     {
+        $this->authorize('show', $tower);
+
         return new TowerResource($tower);
     }
 
@@ -70,9 +80,12 @@ class TowerController extends Controller
      * @param UpdateRequest $request
      * @param Tower $tower
      * @return TowerResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, Tower $tower)
     {
+        $this->authorize('update', $tower);
+
         $tower = $this->towerRepository->update($tower, $request->all());
 
         return new TowerResource($tower);
@@ -83,9 +96,12 @@ class TowerController extends Controller
      *
      * @param Tower $tower
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(Tower $tower)
     {
+        $this->authorize('destroy', $tower);
+
         $this->towerRepository->delete($tower);
 
         return response()->json(null, 204);
